pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "selenium>=4.35.0",
    "socketio>=0.2.1",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "arrow>=1.3.0",
    "requests>=2.32.5",
    "fire>=0.7.1",
    "schedule>=1.2.2",
    "apscheduler>=3.11.0",
    "numpy>=2.3.2",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
    "ratelimit>=2.2.1",
    "aiohttp>=3.12.15",
    "discord-py>=2.6.2",
    "loguru>=0.7.3",
]
